Rename stores -> store
